+++ pakyow-core/spec/features/operation/instance_variables_spec.rb
@@ -20,4 +20,22 @@ RSpec.describe "accessing values through their instance variable" do
   it "exposes the value" do
     expect(app.operations.foo(values).instance_foo).to eq("foo")
   end
+
+  context "optional value is not passed" do
+    let(:app_def) {
+      Proc.new {
+        operation :foo do
+          optional :foo
+        end
+      }
+    }
+
+    let(:values) {
+      {}
+    }
+
+    it "defines an instance variable" do
+      expect(app.operations.foo.instance_variable_defined?(:@foo)).to be(true)
+    end
+  end
 end